Add delete functionality + book cover art
Proper delete functionality has been added to each book, when the user
presses the X button the book is removed from the myLibrary array and
all the books are rendered again.

A new input has been added to the form where the user may include an
url to an image and it will be saved and shown on the rendered book.